requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
anyio==4.4.0
attrs==24.2.0
azure-core==1.31.0
azure-identity==1.17.1
blinker==1.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.1
Deprecated==1.2.14
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.2
hyperframe==6.0.1
idna==3.8
importlib_metadata==8.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
microsoft-kiota-abstractions==1.3.3
microsoft-kiota-authentication-azure==1.1.0
microsoft-kiota-http==1.3.3
microsoft-kiota-serialization-form==0.1.1
microsoft-kiota-serialization-json==1.3.2
microsoft-kiota-serialization-multipart==0.1.0
microsoft-kiota-serialization-text==1.0.0
msal==1.31.0
msal-extensions==1.2.0
msgraph-core==1.1.3
msgraph-sdk==1.7.0
multidict==6.1.0
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
pendulum==3.0.0
pip-autoremove==0.10.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
std-uritemplate==1.0.6
time-machine==2.15.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.4
wrapt==1.16.0
yarl==1.11.1
zipp==3.20.1
